+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.0.2
 
 - Move to github
+- Bump Rx
 
 ## 0.0.1